feat: centralize numeric parsing
